restructuring and changed index.js file names
